no warning at all

+++ lib/utils.cpp
@@ -17,9 +17,9 @@ namespace utils {
 	int read_reverse(
         unsigned char *dst, 
         const unsigned char *src, 
-        size_t dst_size, 
-        size_t src_size, 
-        int *current_index, 
+        ssize_t dst_size, 
+        ssize_t src_size, 
+        ssize_t *current_index, 
         bool strict_size
     ) {
